+++ nodeconductor/structure/apps.py
@@ -157,9 +157,16 @@ class StructureConfig(AppConfig):
                     model.__name__, index),
             )
 
+            structure_signals.resource_imported.connect(
+                handlers.log_resource_imported,
+                sender=model,
+                dispatch_uid='nodeconductor.structure.handlers.log_resource_imported_{}_{}'.format(
+                    model.__name__, index),
+            )
+
         for index, model in enumerate(Resource.get_vm_models()):
             if issubclass(model, CoordinatesMixin):
-                signals.post_save.connect(
+                fsm_signals.post_transition.connect(
                     handlers.detect_vm_coordinates,
                     sender=model,
                     dispatch_uid='nodeconductor.structure.handlers.detect_vm_coordinates_{}_{}'.format(